requirements.txt
Flask==3.0.3
google-cloud-aiplatform==1.71.1
google-generativeai==0.8.3
python-dotenv==1.0.0
flask-cors==5.0.0
opencv-python-headless==4.10.0.84
pillow==11.0.0
numpy==2.1.3
imutils==0.5.4